--- src/auth/auth.module.ts
@@ -1,26 +1,0 @@
-import { Module } from '@nestjs/common';
-import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
-import { AuthRepository } from './repositories/auth.repository';
-import { KakaoStrategy } from './strategies/kakao.strategy';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { PrismaModule } from '../../prisma/prisma.module';
-import { JwtModule } from '@nestjs/jwt';
-
-@Module({
-  imports: [
-    ConfigModule,
-    PrismaModule,
-    JwtModule.registerAsync({
-      useFactory: (configService: ConfigService) => ({
-        secret: configService.get<string>('JWT_SECRET'),
-        signOptions: { expiresIn: '1d' },
-      }),
-      inject: [ConfigService],
-    }),
-  ],
-  controllers: [AuthController],
-  providers: [AuthService, AuthRepository, KakaoStrategy],
-  exports: [AuthService, AuthRepository],
-})
-export class AuthModule {}
