+++ src/layer/modules/auth.module.ts
@@ -1,0 +1,29 @@
+import { Module } from '@nestjs/common';
+import { AuthController } from '../controllers/auth.controller';
+import { AuthService } from '../services/auth.service';
+import { AuthRepository } from '../repositories/auth.repository';
+import { KakaoStrategy } from '../strategies/kakao.strategy';
+import { JwtStrategy } from '../strategies/jwt.strategy';
+import { PassportModule } from '@nestjs/passport';
+import { ConfigModule, ConfigService } from '@nestjs/config';
+import { PrismaModule } from '../../../prisma/prisma.module';
+import { JwtModule } from '@nestjs/jwt';
+
+@Module({
+  imports: [
+    PassportModule.register({ defaultStrategy: 'jwt', session: false }),
+    ConfigModule,
+    PrismaModule,
+    JwtModule.registerAsync({
+      useFactory: (configService: ConfigService) => ({
+        secret: configService.get<string>('JWT_SECRET') || 'default-secret',
+        signOptions: { expiresIn: '1d' },
+      }),
+      inject: [ConfigService],
+    }),
+  ],
+  controllers: [AuthController],
+  providers: [AuthService, AuthRepository, KakaoStrategy, JwtStrategy],
+  exports: [AuthService, AuthRepository],
+})
+export class AuthModule {}
